# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CLARK Gateway",
    "version": "3.7.6",
    "description": "The API gateway for all of the CLARK microservices",
    "license": { "name": "ISC", "url": "https://www.isc.org/licenses/" },
    "contact": {
      "name": "SecurEd Inc.",
      "url": "https://secured.team/",
      "email": "info@secured.team"
    },
    "x-logo": {
      "url": "https://clark.center/assets/images/logo.png",
      "altText": "CLARK Logo"
    },
    "x-tagGroups": [
      { "name": "Library Service", "tags": ["Library Service"] },
      { "name": "Notification Service", "tags": ["Notification Service"] },
      { "name": "Feature Service", "tags": ["Feature Service"] },
      { "name": "Rating Service", "tags": ["Rating Service"] },
      { "name": "Outcome Service", "tags": ["Outcome Service"] },
      {
        "name": "Learning Object Service",
        "tags": ["Learning Object Service"]
      },
      { "name": "User Service", "tags": ["User Service"] },
      { "name": "Utility Service", "tags": ["Utility Service"] },
      {
        "name": "Standard Guideline Service",
        "tags": ["Standard Guideline Service"]
      }
    ]
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Development" },
    { "url": "https://api-gateway.clark.center", "description": "Production" }
  ],
  "paths": {
...
    "/learning-objects/{learningObjectId}/learning-outcomes": {
      "post": {
        "description": "Adds a learning outcome",
        "tags": ["Learning Object Service"],
        "parameters": [
          {
            "in": "path",
            "name": "learningObjectId",
            "schema": { "type": "string" },
            "required": true,
            "description": "The id of the learning object"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Partial outcome",
                "$ref": "#/components/schemas/Outcome"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "description": "The id of the new outcome",
                      "example": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST - Request body missing required information to make a outcome"
          },
          "401": { "description": "UNAUTHENTICATED - User not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User is not the author or privileged user"
          },
          "404": { "description": "NOT FOUND - Outcome not found" }
        }
      }
    },
    "/learning-objects/{learningObjectId}/learning-outcomes/{outcomeId}": {
      "patch": {
        "description": "Updates a outcome",
        "tags": ["Learning Object Service"],
        "parameters": [
          {
            "in": "path",
            "name": "learningObjectId",
            "schema": { "type": "string" },
            "required": true,
            "description": "The id of the learning object"
          },
          {
            "in": "path",
            "name": "outcomeId",
            "schema": { "type": "string" },
            "required": true,
            "description": "The id of the outcome"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Partial outcome",
                "$ref": "#/components/schemas/Outcome"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Outcome"
                }
              }
            }
          },
          "400": {
            "description": "BAD REQUEST - Request body missing required information to update the outcome"
          },
          "401": { "description": "UNAUTHENTICATED - User not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User is not the author or privileged user"
          },
          "404": { "description": "NOT FOUND - Outcome or object not found" }
        }
      },
      "delete": {
        "description": "Deletes an outcome",
        "tags": ["Learning Object Service"],
        "parameters": [
          {
            "in": "path",
            "name": "learningObjectId",
            "schema": { "type": "string" },
            "required": true,
            "description": "The id of the learning object"
          },
          {
            "in": "path",
            "name": "outcomeId",
            "schema": { "type": "string" },
            "required": true,
            "description": "The id of the outcome"
          }
        ],
        "responses": {
          "204": { "description": "NO CONTENT" },
          "401": { "description": "UNAUTHENTICATED - User not logged in" },
          "403": {
            "description": "UNAUTHORIZED - User is not the author or privileged user"
          },
          "404": { "description": "NOT FOUND - Outcome or object not found" }
        }
      }
    },
    "/users/{username}/learning-objects/{learningObjectId}/outcomes": {
      "get": {
        "description": "Gets a object's outcomes",
        "tags": ["Learning Object Service"],
        "parameters": [
          {
            "in": "path",
            "name": "learningObjectId",
            "schema": { "type": "string" },
            "required": true,
            "description": "The id of the learning object"
          },
          {
            "in": "path",
            "name": "username",
            "schema": { "type": "string" },
            "required": true,
            "description": "The username of the author"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Outcome" }
                }
              }
            }
          },
          "404": { "description": "NOT FOUND - Learning object not found" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Outcome": {
        "properties": {
          "id": {
            "type": "string",
            "example": 0,
            "required": true,
            "description": "The id of the outcome"
          },
          "bloom": {
            "type": "string",
            "example": "Remember",
            "required": true,
            "description": "The outcome's bloom type"
          },
          "verb": {
            "type": "string",
            "example": "Classify",
            "required": true,
            "description": "The outcome's verb"
          },
          "text": {
            "type": "string",
            "example": "that you can't divide by 0",
            "required": true,
            "description": "The outcome's text"
          },
          "mappings": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/StandardOutcome" }
          }
        }
      },
      "StandardOutcome": {
        "properties": {
          "id": {
            "type": "string",
            "example": 0,
            "required": true,
            "description": "The id of the standard outcome"
          },
          "author": {
            "type": "string",
            "example": "CS2013",
            "required": true,
            "description": "The organization who created the outcome"
          },
          "source": {
            "type": "string",
            "example": "CS2013",
            "required": true,
            "description": "The shortened organization who created the outcome"
          },
          "name": {
            "type": "string",
            "example": "Algorithmic Strategies",
            "required": true,
            "description": "The type of the outcome"
          },
          "date": {
            "type": "string",
            "example": 2013,
            "required": true,
            "description": "The year the outcome was made"
          },
          "outcome": {
            "type": "string",
            "example": "This is a outcome",
            "required": true,
            "description": "The outcome text"
          }
        }
      },
...
    }
  },
...
}